requirements.txt
# Core Trading
ib_insync==0.9.86
redis==5.0.1
msgpack==1.0.7

# Data & Numerics
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
statsmodels==0.14.1
scikit-learn==1.3.2

# Models & Config
pydantic==2.5.3
pyyaml==6.0.1

# Async
uvloop==0.19.0;sys_platform!='win32'

# Monitoring & API
prometheus-client==0.19.0
fastapi==0.108.0
uvicorn==0.25.0

# Backtesting & Reporting
pyfolio-reloaded==0.9.5
matplotlib==3.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Logging
python-json-logger==2.0.7
